♻️ refactor: EventService를 Command, Query Service로 분리
