+++ src/main/java/com/grepp/spring/app/controller/api/event/EventController.java
@@ -6,7 +6,8 @@ import com.grepp.spring.app.controller.api.event.payload.response.AllTimeSchedul
 import com.grepp.spring.app.controller.api.event.payload.response.CreateEventResponse;
 import com.grepp.spring.app.controller.api.event.payload.response.ScheduleResultResponse;
 import com.grepp.spring.app.controller.api.event.payload.response.ShowEventResponse;
-import com.grepp.spring.app.model.event.service.EventService;
+import com.grepp.spring.app.model.event.service.EventCommandService;
+import com.grepp.spring.app.model.event.service.EventQueryService;
 import com.grepp.spring.infra.error.exceptions.event.EventAuthenticationException;
 import com.grepp.spring.infra.response.ApiResponse;
 import com.grepp.spring.infra.response.EventErrorCode;
@@ -27,7 +28,8 @@ import org.springframework.web.bind.annotation.*;
 @Slf4j
 public class EventController {
 
-    private final EventService eventService;
+    private final EventQueryService eventQueryService;
+    private final EventCommandService eventCommandService;
 
     @PostMapping
     @Operation(summary = "이벤트 생성", description = "그룹 이벤트 또는 일회성 이벤트를 생성합니다.")
@@ -35,7 +37,7 @@ public class EventController {
 
         String currentMemberId = extractCurrentMemberId();
 
-        CreateEventResponse response = eventService.createEvent(request, currentMemberId);
+        CreateEventResponse response = eventCommandService.createEvent(request, currentMemberId);
 
         return ResponseEntity.status(200)
             .body(ApiResponse.success("이벤트가 성공적으로 생성되었습니다.", response));
@@ -49,7 +51,7 @@ public class EventController {
 
         String currentMemberId = extractCurrentMemberId();
 
-        ShowEventResponse response = eventService.getEvent(eventId, currentMemberId);
+        ShowEventResponse response = eventQueryService.getEvent(eventId, currentMemberId);
 
         return ResponseEntity.ok(ApiResponse.success("이벤트 조회가 성공적으로 완료되었습니다.", response));
 
@@ -64,7 +66,7 @@ public class EventController {
 
         String currentMemberId = extractCurrentMemberId();
 
-        eventService.joinEvent(eventId, groupId, currentMemberId);
+        eventCommandService.joinEvent(eventId, groupId, currentMemberId);
 
         return ResponseEntity.ok(ApiResponse.success("이벤트에 성공적으로 참여했습니다."));
 
@@ -79,7 +81,7 @@ public class EventController {
 
         String currentMemberId = extractCurrentMemberId();
 
-        eventService.createOrUpdateMyTime(request, eventId, currentMemberId);
+        eventCommandService.createOrUpdateMyTime(request, eventId, currentMemberId);
 
         return ResponseEntity.ok(ApiResponse.success("개인 일정이 성공적으로 생성/수정되었습니다."));
 
@@ -92,7 +94,7 @@ public class EventController {
 
         String currentMemberId = extractCurrentMemberId();
 
-        AllTimeScheduleResponse response = eventService.getAllTimeSchedules(eventId, currentMemberId);
+        AllTimeScheduleResponse response = eventQueryService.getAllTimeSchedules(eventId, currentMemberId);
 
         return ResponseEntity.ok(ApiResponse.success(response));
 
@@ -105,7 +107,7 @@ public class EventController {
 
         String currentMemberId = extractCurrentMemberId();
 
-        eventService.completeMyTime(eventId, currentMemberId);
+        eventCommandService.completeMyTime(eventId, currentMemberId);
 
         return ResponseEntity.ok(ApiResponse.success("개인 일정이 성공적으로 확정되었습니다."));
 
@@ -119,7 +121,7 @@ public class EventController {
 
         String currentMemberId = extractCurrentMemberId();
 
-        eventService.createScheduleResult(eventId, currentMemberId);
+        eventCommandService.createScheduleResult(eventId, currentMemberId);
 
         return ResponseEntity.ok(ApiResponse.success("일정 조율 결과가 성공적으로 생성되었습니다."));
 
@@ -132,7 +134,7 @@ public class EventController {
 
         String currentMemberId = extractCurrentMemberId();
 
-        ScheduleResultResponse response = eventService.getScheduleResult(eventId, currentMemberId);
+        ScheduleResultResponse response = eventQueryService.getScheduleResult(eventId, currentMemberId);
 
         return ResponseEntity.ok(ApiResponse.success(response));
