+++ src/main/java/com/grepp/spring/app/model/event/service/EventQueryService.java
@@ -1,36 +1,26 @@
 package com.grepp.spring.app.model.event.service;
 
-import com.grepp.spring.app.controller.api.event.payload.request.CreateEventRequest;
-import com.grepp.spring.app.controller.api.event.payload.request.MyTimeScheduleRequest;
 import com.grepp.spring.app.controller.api.event.payload.response.AllTimeScheduleResponse;
-import com.grepp.spring.app.controller.api.event.payload.response.CreateEventResponse;
 import com.grepp.spring.app.controller.api.event.payload.response.ScheduleResultResponse;
 import com.grepp.spring.app.controller.api.event.payload.response.ShowEventResponse;
-import com.grepp.spring.app.model.event.code.Role;
-import com.grepp.spring.app.model.event.dto.*;
+import com.grepp.spring.app.model.event.dto.AllTimeScheduleDto;
+import com.grepp.spring.app.model.event.dto.ScheduleResultDto;
 import com.grepp.spring.app.model.event.entity.CandidateDate;
 import com.grepp.spring.app.model.event.entity.Event;
 import com.grepp.spring.app.model.event.entity.EventMember;
 import com.grepp.spring.app.model.event.entity.TempSchedule;
-import com.grepp.spring.app.model.event.factory.EventCreationStrategyFactory;
 import com.grepp.spring.app.model.event.repository.CandidateDateRepository;
 import com.grepp.spring.app.model.event.repository.EventMemberRepository;
 import com.grepp.spring.app.model.event.repository.EventRepository;
 import com.grepp.spring.app.model.event.repository.TempScheduleRepository;
-import com.grepp.spring.app.model.event.strategy.EventCreationStrategy;
-import com.grepp.spring.app.model.group.code.GroupRole;
-import com.grepp.spring.app.model.group.entity.Group;
-import com.grepp.spring.app.model.group.entity.GroupMember;
-import com.grepp.spring.app.model.group.repository.GroupMemberRepository;
-import com.grepp.spring.app.model.group.repository.GroupRepository;
-import com.grepp.spring.app.model.member.entity.Member;
-import com.grepp.spring.app.model.member.repository.MemberRepository;
 import com.grepp.spring.app.model.schedule.code.ScheduleStatus;
 import com.grepp.spring.app.model.schedule.entity.Schedule;
 import com.grepp.spring.app.model.schedule.entity.ScheduleMember;
 import com.grepp.spring.app.model.schedule.repository.ScheduleMemberQueryRepository;
 import com.grepp.spring.app.model.schedule.repository.ScheduleQueryRepository;
-import com.grepp.spring.infra.error.exceptions.event.*;
+import com.grepp.spring.infra.error.exceptions.event.EventNotFoundException;
+import com.grepp.spring.infra.error.exceptions.event.NotEventMemberException;
+import com.grepp.spring.infra.error.exceptions.event.ScheduleResultNotFoundException;
 import com.grepp.spring.infra.response.EventErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -46,38 +36,15 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 @Slf4j
-public class EventService {
+public class EventQueryService {
 
     private final EventRepository eventRepository;
     private final EventMemberRepository eventMemberRepository;
     private final CandidateDateRepository candidateDateRepository;
-    private final MemberRepository memberRepository;
-    private final GroupRepository groupRepository;
-    private final GroupMemberRepository groupMemberRepository;
     private final TempScheduleRepository tempScheduleRepository;
-    private final EventScheduleResultService eventScheduleResultService;
     private final ScheduleQueryRepository scheduleQueryRepository;
     private final ScheduleMemberQueryRepository scheduleMemberQueryRepository;
 
-    private final EventCreationStrategyFactory strategyFactory;
-
-    @Transactional
-    public CreateEventResponse createEvent(CreateEventRequest webRequest, String currentMemberId) {
-        CreateEventDto serviceRequest = CreateEventDto.toDto(webRequest, currentMemberId);
-
-        EventCreationStrategy strategy = strategyFactory.getStrategy(serviceRequest.getGroupId());
-        Event event = strategy.createEvent(serviceRequest, currentMemberId);
-
-        CreateEventResponse response = new CreateEventResponse(
-            event.getId(),
-            event.getTitle(),
-            event.getGroup().getId()
-        );
-
-        return response;
-    }
-
-    @Transactional(readOnly = true)
     public ShowEventResponse getEvent(Long eventId, String currentMemberId) {
         Event event = eventRepository.findById(eventId)
             .orElseThrow(() -> new EventNotFoundException(EventErrorCode.EVENT_NOT_FOUND));
@@ -95,112 +62,6 @@ public class EventService {
         return response;
     }
 
-    @Transactional
-    public void joinEvent(Long eventId, Long groupId, String currentMemberId) {
-
-        JoinEventDto dto = JoinEventDto.toDto(eventId, currentMemberId);
-
-        Event event = eventRepository.findById(dto.getEventId())
-            .orElseThrow(() -> new EventNotFoundException(EventErrorCode.EVENT_NOT_FOUND));
-
-        Member member = memberRepository.findById(dto.getMemberId())
-            .orElseThrow(() -> new EventNotFoundException(EventErrorCode.MEMBER_NOT_FOUND));
-
-        if (eventMemberRepository.existsByEventIdAndMemberId(dto.getEventId(), dto.getMemberId())) {
-            throw new AlreadyJoinedEventException(EventErrorCode.ALREADY_JOINED_EVENT);
-        }
-
-        Long currentMemberCount = eventMemberRepository.countByEventId(dto.getEventId());
-        event.validateCapacity(currentMemberCount);
-
-        addUserToGroup(groupId, dto.getMemberId());
-
-        EventMemberDto memberDto = EventMemberDto.toDto(dto.getEventId(), dto.getMemberId(), Role.ROLE_MEMBER);
-        createEventMember(memberDto);
-
-    }
-
-    private void addUserToGroup(Long groupId, String memberId) {
-        try {
-            Group group = groupRepository.findById(groupId)
-                .orElseThrow(() -> new EventNotFoundException(EventErrorCode.GROUP_NOT_FOUND));
-
-            Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new EventNotFoundException(EventErrorCode.MEMBER_NOT_FOUND));
-
-            Optional<GroupMember> existingGroupMember = groupMemberRepository
-                .findByGroupIdAndMemberId(groupId, memberId);
-
-            if (existingGroupMember.isPresent()) {
-                return;
-            }
-
-            GroupMember groupMember = new GroupMember();
-            groupMember.setGroup(group);
-            groupMember.setMember(member);
-            groupMember.setRole(GroupRole.GROUP_MEMBER);
-            groupMember.setGroupAdmin(false);
-
-            groupMemberRepository.save(groupMember);
-
-        } catch (EventNotFoundException e) {
-            throw e;
-        } catch (Exception e) {
-            throw new InvalidEventDataException(EventErrorCode.INVALID_EVENT_DATA);
-        }
-    }
-
-    @Transactional
-    public void createEventMember(EventMemberDto dto) {
-        Event event = eventRepository.findById(dto.getEventId())
-            .orElseThrow(() -> new EventNotFoundException(EventErrorCode.EVENT_NOT_FOUND));
-
-        Member member = memberRepository.findById(dto.getMemberId())
-            .orElseThrow(() -> new EventNotFoundException(EventErrorCode.MEMBER_NOT_FOUND));
-
-        EventMember eventMember = EventMemberDto.toEntity(dto, event, member);
-        eventMemberRepository.save(eventMember);
-    }
-
-    @Transactional
-    public void createOrUpdateMyTime(MyTimeScheduleRequest request, Long eventId, String currentMemberId) {
-        MyTimeScheduleDto dto = MyTimeScheduleDto.toDto(request, eventId, currentMemberId);
-
-        Event event = eventRepository.findById(dto.getEventId())
-            .orElseThrow(() -> new EventNotFoundException(EventErrorCode.EVENT_NOT_FOUND));
-
-        EventMember eventMember = eventMemberRepository.findByEventIdAndMemberIdAndActivatedTrue(dto.getEventId(), dto.getMemberId())
-            .orElseThrow(() -> new NotEventMemberException(EventErrorCode.NOT_EVENT_MEMBER));
-
-        if (eventMember.getConfirmed()) {
-            throw new AlreadyCompletedScheduleException(EventErrorCode.ALREADY_COMPLETED_SCHEDULE);
-        }
-
-        for (MyTimeScheduleDto.DailyTimeSlotDto slot : dto.getDailyTimeSlots()) {
-            updateOrCreateTempSchedule(eventMember, slot);
-        }
-    }
-
-    private void updateOrCreateTempSchedule(EventMember eventMember, MyTimeScheduleDto.DailyTimeSlotDto slot) {
-        LocalDate date = slot.getDate();
-
-        Optional<TempSchedule> existingSchedule = tempScheduleRepository
-            .findByEventMemberIdAndDateAndActivatedTrue(eventMember.getId(), date);
-
-        if (existingSchedule.isPresent()) {
-            TempSchedule schedule = existingSchedule.get();
-            Long currentTimeBit = schedule.getTimeBit();
-            Long newTimeBit = currentTimeBit ^ slot.getTimeBitAsLong();
-
-            schedule.setTimeBit(newTimeBit);
-            tempScheduleRepository.save(schedule);
-        } else {
-            TempSchedule newSchedule = MyTimeScheduleDto.DailyTimeSlotDto.toEntity(slot, eventMember);
-            tempScheduleRepository.save(newSchedule);
-        }
-    }
-
-    @Transactional(readOnly = true)
     public AllTimeScheduleResponse getAllTimeSchedules(Long eventId, String currentMemberId) {
         Event event = eventRepository.findById(eventId)
             .orElseThrow(() -> new EventNotFoundException(EventErrorCode.EVENT_NOT_FOUND));
@@ -358,40 +219,6 @@ public class EventService {
             .build();
     }
 
-    @Transactional
-    public void completeMyTime(Long eventId, String currentMemberId) {
-        JoinEventDto dto = JoinEventDto.toDto(eventId, currentMemberId);
-
-        Event event = eventRepository.findById(dto.getEventId())
-            .orElseThrow(() -> new EventNotFoundException(EventErrorCode.EVENT_NOT_FOUND));
-
-        EventMember eventMember = eventMemberRepository
-            .findByEventIdAndMemberIdAndActivatedTrue(dto.getEventId(), dto.getMemberId())
-            .orElseThrow(() -> new NotEventMemberException(EventErrorCode.NOT_EVENT_MEMBER));
-
-        List<TempSchedule> schedules = tempScheduleRepository
-            .findAllByEventMemberIdAndActivatedTrue(eventMember.getId());
-
-        if (schedules.isEmpty()) {
-            throw new InvalidEventDataException(EventErrorCode.CANNOT_COMPLETE_EMPTY_SCHEDULE);
-        }
-
-        eventMember.confirmScheduleOrThrow();
-        eventMemberRepository.save(eventMember);
-    }
-
-    @Transactional
-    public void createScheduleResult(Long eventId, String currentMemberId) {
-        Event event = eventRepository.findById(eventId)
-            .orElseThrow(() -> new EventNotFoundException(EventErrorCode.EVENT_NOT_FOUND));
-
-        EventMember eventMember = eventMemberRepository.findByEventIdAndMemberIdAndActivatedTrue(eventId, currentMemberId)
-            .orElseThrow(() -> new NotEventMemberException(EventErrorCode.NOT_EVENT_MEMBER));
-
-        eventScheduleResultService.createScheduleRecommendations(eventId);
-    }
-
-    @Transactional(readOnly = true)
     public ScheduleResultResponse getScheduleResult(Long eventId, String currentMemberId) {
         Event event = eventRepository.findById(eventId)
             .orElseThrow(() -> new EventNotFoundException(EventErrorCode.EVENT_NOT_FOUND));
